# backend/specs/openapi.yaml
openapi: 3.1.0
info:
  title: NarrativeOS API
  version: 0.3.0
paths:
...
  /v1/reader/library/worlds/{world_id}:
    get:
      tags:
      - reader
      summary: World Detail
      operationId: world_detail_v1_reader_library_worlds__world_id__get
      parameters:
      - name: world_id
        in: path
        required: true
        schema:
          type: string
          title: World Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response World Detail V1 Reader Library Worlds  World Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/sessions:
    post:
      tags:
      - reader
      summary: Create Reader Session
      operationId: create_reader_session_v1_reader_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReaderSessionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Reader Session V1 Reader Sessions Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/continue:
    post:
      tags:
      - reader
      summary: Continue Reader Story
      operationId: continue_reader_story_v1_reader_continue_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinueReaderRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Continue Reader Story V1 Reader Continue Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/entitlements:
    get:
      tags:
      - reader
      summary: Reader Entitlements
      operationId: reader_entitlements_v1_reader_entitlements_get
      parameters:
      - name: reader_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
      - name: world_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: World Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Entitlements V1 Reader Entitlements Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/subscription:
    get:
      tags:
      - reader
      summary: Reader Subscription
      operationId: reader_subscription_v1_reader_subscription_get
      parameters:
      - name: reader_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Subscription V1 Reader Subscription Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/entitlements/grant:
    post:
      tags:
      - reader
      summary: Grant Reader Entitlement
      operationId: grant_reader_entitlement_v1_reader_entitlements_grant_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantEntitlementRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Grant Reader Entitlement V1 Reader Entitlements Grant
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/checkout/start:
    post:
      tags:
      - reader
      summary: Start Checkout
      operationId: start_checkout_v1_reader_checkout_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartCheckoutRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Start Checkout V1 Reader Checkout Start Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/checkout/{checkout_session_id}/status:
    get:
      tags:
      - reader
      summary: Reader Checkout Status
      operationId: reader_checkout_status_v1_reader_checkout__checkout_session_id__status_get
      parameters:
      - name: checkout_session_id
        in: path
        required: true
        schema:
          type: string
          title: Checkout Session Id
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
      - name: reader_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
      - name: include_diagnostics
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Diagnostics
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Checkout Status V1 Reader Checkout  Checkout
                  Session Id  Status Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/checkout/return:
    post:
      tags:
      - reader
      summary: Reader Checkout Return
      operationId: reader_checkout_return_v1_reader_checkout_return_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutReturnRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Reader Checkout Return V1 Reader Checkout Return Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/checkout/provider-callback:
    post:
      tags:
      - reader
      summary: Reader Checkout Provider Callback
      operationId: reader_checkout_provider_callback_v1_reader_checkout_provider_callback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutWebhookRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Reader Checkout Provider Callback V1 Reader Checkout
                  Provider Callback Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/checkout/webhook:
    post:
      tags:
      - reader
      summary: Reader Checkout Webhook
      operationId: reader_checkout_webhook_v1_reader_checkout_webhook_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutWebhookRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Reader Checkout Webhook V1 Reader Checkout Webhook
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/subscription/{account_id}/retry-payment:
    post:
      tags:
      - reader
      summary: Reader Retry Subscription Payment
      operationId: reader_retry_subscription_payment_v1_reader_subscription__account_id__retry_payment_post
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Retry Subscription Payment V1 Reader Subscription  Account
                  Id  Retry Payment Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/subscription/{account_id}/renew:
    post:
      tags:
      - reader
      summary: Reader Renew Subscription
      operationId: reader_renew_subscription_v1_reader_subscription__account_id__renew_post
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Renew Subscription V1 Reader Subscription  Account
                  Id  Renew Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/subscription/{account_id}/cancel:
    post:
      tags:
      - reader
      summary: Reader Cancel Subscription
      operationId: reader_cancel_subscription_v1_reader_subscription__account_id__cancel_post
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: string
          title: Account Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Cancel Subscription V1 Reader Subscription  Account
                  Id  Cancel Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/sessions/{session_id}/replay:
    get:
      tags:
      - reader
      summary: Reader Replay
      operationId: reader_replay_v1_reader_sessions__session_id__replay_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Replay V1 Reader Sessions  Session Id  Replay
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/sessions/{session_id}/prefill:
    get:
      tags:
      - reader
      summary: Reader Prefill
      operationId: reader_prefill_v1_reader_sessions__session_id__prefill_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Prefill V1 Reader Sessions  Session Id  Prefill
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/reader/sessions/{session_id}/quote:
    get:
      tags:
      - reader
      summary: Reader Quote
      operationId: reader_quote_v1_reader_sessions__session_id__quote_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reader Quote V1 Reader Sessions  Session Id  Quote
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/library/worlds/{world_id}:
    get:
      summary: Get Library World
      operationId: get_library_world_v1_library_worlds__world_id__get
      parameters:
      - name: world_id
        in: path
        required: true
        schema:
          type: string
          title: World Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Library World V1 Library Worlds  World Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/sessions:
    post:
      summary: Create Session
      operationId: create_session_v1_sessions_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Session V1 Sessions Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Sessions
      operationId: list_sessions_v1_sessions_get
      parameters:
      - name: world_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: World Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Sessions V1 Sessions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/sessions/{session_id}/replay:
    get:
      summary: Replay Session
      operationId: replay_session_v1_sessions__session_id__replay_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Replay Session V1 Sessions  Session Id  Replay Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/sessions/{session_id}/prefill:
    get:
      summary: Session Prefill
      operationId: session_prefill_v1_sessions__session_id__prefill_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Session Prefill V1 Sessions  Session Id  Prefill Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    ContinueReaderRequest:
      properties:
        session_id:
          type: string
          title: Session Id
        choice_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Choice Id
        freeform_intent:
          anyOf:
          - type: string
          - type: 'null'
          title: Freeform Intent
        account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
        reader_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
      type: object
      required:
      - session_id
      title: ContinueReaderRequest
    CreateReaderSessionRequest:
      properties:
        world_id:
          type: string
          title: World Id
        account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
        reader_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
      type: object
      required:
      - world_id
      title: CreateReaderSessionRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StartCheckoutRequest:
      properties:
        account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
        reader_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
        tier_id:
          type: string
          title: Tier Id
        provider:
          type: string
          title: Provider
          default: web_stub
      type: object
      required:
      - tier_id
      title: StartCheckoutRequest
...